requirements.txt
switch2bill-common==2.3
linaro-django-pagination==2.0.2
git+git://github.com/Star2Billing/django-notification
